feat: creating plugin for mode toggle menu component
