+++ packages/react-tinacms-editor/src/components/MarkdownEditor/Menubar/index.tsx
@@ -16,60 +16,48 @@ limitations under the License.
 
 */
 
-import React from 'react'
-import { useState, useRef, useEffect, useLayoutEffect } from 'react'
+import React, { useState, useRef, useEffect, useLayoutEffect } from 'react'
 
-import { TablePopups } from '../../plugins/Table/Popup'
-import { WysiwygMenu as BlockMenu } from '../../plugins/Block'
-import { WysiwygMenu as CodeBlockMenu } from '../../plugins/CodeBlock'
-import { WysiwygMenu as HistoryMenu } from '../../plugins/History'
-import { WysiwygMenu as InlineMenu } from '../../plugins/Inline'
-import { WysiwygMenu as ListMenu } from '../../plugins/List'
-import { WysiwygMenu as QuoteMenu } from '../../plugins/Blockquote'
-import { WysiwygMenu as TableMenu } from '../../plugins/Table'
-import {
-  ImageEdit as ImageEditPopup,
-  Loader as ImageLoader,
-  WysiwygMenu as ImageMenu,
-} from '../../plugins/Image'
-import {
-  LinkForm as LinkFormPopup,
-  WysiwygMenu as LinkMenu,
-} from '../../plugins/Link'
-
-import { useEditorStateContext } from '../../context/editorState'
-import { MenuPortalProvider } from '../../context/MenuPortal'
-import { EditorModeMenu } from '../EditorModeMenu'
-
+import { MarkdownMenuItem as BlockMenu } from '../../../plugins/Block'
+import { MarkdownMenuItem as InlineMenu } from '../../../plugins/Inline'
+import { MarkdownMenuItem as LinkMenu } from '../../../plugins/Link'
+import { MarkdownMenuItem as ImageMenu } from '../../../plugins/Image'
+import { MarkdownMenuItem as TableMenu } from '../../../plugins/Table'
+import { MarkdownMenuItem as QuoteMenu } from '../../../plugins/Blockquote'
+import { MarkdownMenuItem as CodeBlockMenu } from '../../../plugins/CodeBlock'
+import { MarkdownMenuItem as ListMenu } from '../../../plugins/List'
+import { MarkdownMenuItem as HistoryMenu } from '../../../plugins/History'
+import { MenuPortalProvider } from '../../../context/MenuPortal'
+import { ImageProps } from '../../../types'
 import {
   MenuPlaceholder,
   MenuWrapper,
   MenuContainer,
-} from '../MenuHelpers/styledComponents'
+} from '../../MenuHelpers/styledComponents'
+import { EditorModeMenu } from '../../EditorModeMenu'
 
 interface Props {
   sticky?: boolean | string
-  uploadImages?: (files: File[]) => Promise<string[]>
-  toggleEditorMode?: () => void
+  toggleEditorMode: () => void
+  imageProps?: ImageProps
 }
 
 export const Menubar = ({
   sticky = true,
-  uploadImages,
   toggleEditorMode,
+  imageProps,
 }: Props) => {
   const [menuFixed, setMenuFixed] = useState(false)
   const isBrowser = typeof window !== `undefined`
   const menuRef = useRef<HTMLDivElement>(null)
   const [menuBoundingBox, setMenuBoundingBox] = useState<any>(null)
   const menuFixedTopOffset = typeof sticky === 'string' ? sticky : '0'
-  const { editorView } = useEditorStateContext()
 
   useEffect(() => {
     if (menuRef.current && sticky) {
       setMenuBoundingBox(menuRef.current.getBoundingClientRect())
     }
-  }, [menuRef, editorView])
+  }, [menuRef])
 
   useLayoutEffect(() => {
     if (!isBrowser || !menuRef.current || !sticky) {
@@ -108,12 +96,10 @@ export const Menubar = ({
     }
   }, [menuRef, menuBoundingBox])
 
-  const preventProsemirrorFocusLoss = React.useCallback((e: any) => {
+  const stopEvent = React.useCallback((e: any) => {
     e.stopPropagation()
     e.preventDefault()
   }, [])
-
-  if (!editorView) return null
 
   return (
     <>
@@ -127,28 +113,20 @@ export const Menubar = ({
         ref={menuRef}
       >
         <MenuPortalProvider>
-          <MenuContainer onMouseDown={preventProsemirrorFocusLoss}>
+          <MenuContainer onMouseDown={stopEvent}>
             <BlockMenu />
             <InlineMenu />
             <LinkMenu />
-            <ImageMenu uploadImages={uploadImages} />
+            {imageProps && <ImageMenu />}
             <TableMenu />
             <QuoteMenu />
             <CodeBlockMenu />
             <ListMenu />
             <HistoryMenu />
-            {toggleEditorMode && (
-              <EditorModeMenu toggleEditorMode={toggleEditorMode} />
-            )}
+            <EditorModeMenu toggleEditorMode={toggleEditorMode} />
           </MenuContainer>
         </MenuPortalProvider>
       </MenuWrapper>
-      <TablePopups />
-      <ImageEditPopup />
-      <LinkFormPopup />
-      <ImageLoader />
     </>
   )
 }
-
-// todo: sub-menus to return null if schema does not have related type of node / mark.
