+++ packages/react-tinacms-editor/src/plugins/Blockquote/Menu/MenuItem.tsx
@@ -40,7 +40,7 @@ function wrapInBlockquote(state: EditorState, dispatch: any) {
   return wrapIn(state.schema.nodes.blockquote)(state, dispatch)
 }
 
-export const WysiwygMenu = commandControl(
+export const MenuItem = commandControl(
   wrapInBlockquote,
   QuoteIcon,
   'Blockquote',
