+++ packages/react-tinacms-editor/src/plugins/History/Menu/MenuItem.tsx
@@ -24,7 +24,7 @@ import { RedoIcon, UndoIcon } from '@tinacms/icons'
 import { useEditorStateContext } from '../../../context/editorState'
 import { MenuButton } from '../../../components/MenuHelpers'
 
-export const WysiwygMenu = () => (
+export const MenuItem = () => (
   <>
     <UndoControl />
     <RedoControl />
