+++ packages/react-tinacms-editor/src/plugins/Inline/Menu/MarkdownMenuItem.tsx
@@ -19,30 +19,18 @@ limitations under the License.
 import * as React from 'react'
 import { BoldIcon, ItalicIcon, StrikethroughIcon } from '@tinacms/icons'
 
-import { markControl } from '../../../components/MenuHelpers'
+import { MenuButton } from '../../../components/MenuHelpers'
 
-export const WysiwygMenu = () => (
+export const MarkdownMenuItem = () => (
   <>
-    <BoldControl />
-    <ItalicControl />
-    <StrikeControl />
+    <MenuButton data-tooltip="Bold" data-side="top" disabled>
+      <BoldIcon />
+    </MenuButton>
+    <MenuButton data-tooltip="Italic" data-side="top" disabled>
+      <ItalicIcon />
+    </MenuButton>
+    <MenuButton data-tooltip="Strike" data-side="top" disabled>
+      <StrikethroughIcon />
+    </MenuButton>
   </>
 )
-
-const BoldControl = markControl({
-  mark: 'strong',
-  Icon: BoldIcon,
-  tooltip: 'Bold',
-})
-
-const ItalicControl = markControl({
-  mark: 'em',
-  Icon: ItalicIcon,
-  tooltip: 'Italic',
-})
-
-const StrikeControl = markControl({
-  mark: 'strike',
-  Icon: StrikethroughIcon,
-  tooltip: 'Strike',
-})
