+++ packages/react-tinacms-editor/src/plugins/List/Menu/MarkdownMenuItem.tsx
@@ -16,13 +16,26 @@ limitations under the License.
 
 */
 
-import React from 'react'
-import { LinkIcon } from '@tinacms/icons'
+import * as React from 'react'
+import { UnorderedListIcon, OrderedListIcon } from '@tinacms/icons'
 
 import { MenuButton } from '../../../components/MenuHelpers'
 
-export const MarkdownMenu = () => (
-  <MenuButton data-tooltip="Link" data-side="top" disabled>
-    <LinkIcon />
+export const MarkdownMenuItem = (props: any) => (
+  <>
+    <BulletList {...props} />
+    <OrderedList {...props} />
+  </>
+)
+
+const BulletList = () => (
+  <MenuButton data-tooltip="Unordered List" disabled>
+    <UnorderedListIcon />
+  </MenuButton>
+)
+
+const OrderedList = () => (
+  <MenuButton data-tooltip="Ordered List" disabled>
+    <OrderedListIcon />
   </MenuButton>
 )
